+++ tests/rest/test_connection.py
@@ -2331,18 +2331,18 @@ def test_load_result_filters(requests_mock):
 
 class TestLoadStac:
     def test_basic(self, con120):
-        cube = con120.load_stac("https://provide.test/dataset")
+        cube = con120.load_stac("https://provider.test/dataset")
         assert cube.flat_graph() == {
             "loadstac1": {
                 "process_id": "load_stac",
-                "arguments": {"url": "https://provide.test/dataset"},
+                "arguments": {"url": "https://provider.test/dataset"},
                 "result": True,
             }
         }
 
     def test_extents(self, con120):
         cube = con120.load_stac(
-            "https://provide.test/dataset",
+            "https://provider.test/dataset",
             spatial_extent={"west": 1, "south": 2, "east": 3, "north": 4},
             temporal_extent=["2023-05-10", "2023-06-01"],
             bands=["B02", "B03"],
@@ -2351,10 +2351,33 @@ class TestLoadStac:
             "loadstac1": {
                 "process_id": "load_stac",
                 "arguments": {
-                    "url": "https://provide.test/dataset",
+                    "url": "https://provider.test/dataset",
                     "spatial_extent": {"east": 3, "north": 4, "south": 2, "west": 1},
                     "temporal_extent": ["2023-05-10", "2023-06-01"],
                     "bands": ["B02", "B03"],
+                },
+                "result": True,
+            }
+        }
+
+    def test_properties(self, con120):
+        cube = con120.load_stac("https://provider.test/dataset", properties={"platform": lambda p: p == "S2A"})
+        assert cube.flat_graph() == {
+            "loadstac1": {
+                "process_id": "load_stac",
+                "arguments": {
+                    "url": "https://provider.test/dataset",
+                    "properties": {
+                        "platform": {
+                            "process_graph": {
+                                "eq1": {
+                                    "arguments": {"x": {"from_parameter": "value"}, "y": "S2A"},
+                                    "process_id": "eq",
+                                    "result": True,
+                                }
+                            }
+                        }
+                    },
                 },
                 "result": True,
             }